Cleaned up some files, activated git ignore file, remove _site
_site should not be kept under version control, as it is only the
built raw site, not the source content.

Renamed _gitignore to .gitignore so git would use it.

Removed CNAME as we're not hosting this site using github or gitlab
pages.

Fixed pagination icons on the blog. The _include/pagination.html was
copied directly into the blog/index.html file, but the actual icon
changes were made in _include/pagination.html. The include directive
was commented out, resulting in the proper code being added to the
page, but in a comment block and not displayed. Removed the copied
code and uncommented include block.